requirements/common.txt
django==1.8.5
django-suit==0.2.15
djangorestframework==3.2.4
django-rest-swagger==0.3.4
rest_condition